Prevent HAL to rescan path every time setState() is called
The old behavior of the HAL is to define all applicable device
IDs. When all these devices are discovered, the HAL does not
rescan all the device nodes when setState() is called. However,
given that this HAL is to support more and more platforms, and
some of them may not exist on target platform, this behavior
causes the HAL to rescan. This is not necessary.

This patch modifies the scanning function, such that when at least
one device is present, it will not rescan. When one of the node
ceases to exist, the re-scan will be triggered.

Change-Id: Ife119104a77ecde942539853945517686d5bcbe1
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Reviewed-on: https://android.intel.com/228785
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Jovanovic, Radivoje <radivoje.jovanovic@intel.com>