PM: Remove SOC-specific touchboost frequency from Power HAL
Till date we had touchboost_freq assigned in the Power HAL as the
platform-specific HFM. This patch removes this and effectively,
touchboost_freq is initialized to the maximum frequency of the CPU,
as done in the governor. This is done to improve application launch time,
especially in BYT. However, we should keep a watch on Scrolling/Swiping
Customer-specific KPIs with this change.

Change-Id: I9b21a8125ae4268e9aea9f3f4dfc914a76ca2fa4
Signed-off-by: Ashish K <ashish.k@intel.com>
Signed-off-by: Fernand Cuesta <fernand.cuesta@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-12655
Reviewed-on: https://android.intel.com/#/c/306703/
Signed-off-by: Fernand Cuesta <fernand.cuesta@intel.com>
Reviewed-on: https://android.intel.com:443/309503